requirements.txt
flask 
langchain
langchain-community
langchain-google-genai
langgraph
pinecone-client
pinecone
python-dotenv
pydantic
langchain-pinecone
gunicorn